Add: 리팩토링 & 피드백 수정 완료
