+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"/Users/9bin/Desktop/WECODE/westagram-project1/src/index.js":"1","/Users/9bin/Desktop/WECODE/westagram-project1/src/Routes.js":"2","/Users/9bin/Desktop/WECODE/westagram-project1/src/pages/kyubinkim/Login/Login.js":"3","/Users/9bin/Desktop/WECODE/westagram-project1/src/pages/kyubinkim/Main/Main.js":"4","/Users/9bin/Desktop/WECODE/westagram-project1/src/pages/kyubinkim/Main/components/MainComponent.js":"5","/Users/9bin/Desktop/WECODE/westagram-project1/src/pages/kyubinkim/Main/components/HeaderComponent.js":"6","/Users/9bin/Desktop/WECODE/westagram-project1/src/pages/kyubinkim/Main/components/MainLeftComponent.js":"7","/Users/9bin/Desktop/WECODE/westagram-project1/src/pages/kyubinkim/Main/components/MainRightComponent.js":"8"},{"size":185,"mtime":1607581845981,"results":"9","hashOfConfig":"10"},{"size":1500,"mtime":1607580981754,"results":"11","hashOfConfig":"10"},{"size":2567,"mtime":1607670906581,"results":"12","hashOfConfig":"10"},{"size":392,"mtime":1607670782939,"results":"13","hashOfConfig":"10"},{"size":349,"mtime":1607670779547,"results":"14","hashOfConfig":"10"},{"size":1926,"mtime":1607670780840,"results":"15","hashOfConfig":"10"},{"size":4042,"mtime":1607670777962,"results":"16","hashOfConfig":"10"},{"size":2171,"mtime":1607670776405,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"tala7m",{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"/Users/9bin/Desktop/WECODE/westagram-project1/src/index.js",[],["35","36"],"/Users/9bin/Desktop/WECODE/westagram-project1/src/Routes.js",[],"/Users/9bin/Desktop/WECODE/westagram-project1/src/pages/kyubinkim/Login/Login.js",[],"/Users/9bin/Desktop/WECODE/westagram-project1/src/pages/kyubinkim/Main/Main.js",[],"/Users/9bin/Desktop/WECODE/westagram-project1/src/pages/kyubinkim/Main/components/MainComponent.js",[],"/Users/9bin/Desktop/WECODE/westagram-project1/src/pages/kyubinkim/Main/components/HeaderComponent.js",[],"/Users/9bin/Desktop/WECODE/westagram-project1/src/pages/kyubinkim/Main/components/MainLeftComponent.js",[],"/Users/9bin/Desktop/WECODE/westagram-project1/src/pages/kyubinkim/Main/components/MainRightComponent.js",[],{"ruleId":"37","replacedBy":"38"},{"ruleId":"39","replacedBy":"40"},"no-native-reassign",["41"],"no-negated-in-lhs",["42"],"no-global-assign","no-unsafe-negation"]
