+++ src/index.js
@@ -1,5 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import Routes from "./Routes";
+import "./styles/reset.scss";
 
 ReactDOM.render(<Routes />, document.getElementById("root"));
