+++ src/pages/kyubinkim/Main/Main.scss
@@ -1,4 +1,3 @@
-@import "../../../styles/reset.scss";
 @import "../../../styles/common.scss";
 
 .Mainkyubin {
