+++ src/pages/kyubinkim/Main/components/MainComponent.js
@@ -1,0 +1,16 @@
+import React, { Component } from "react";
+import MainLeftComponent from "./MainLeftComponent";
+import MainRightComponent from "./MainRightComponent";
+
+class MainComponent extends Component {
+  render() {
+    return (
+      <main>
+        <MainLeftComponent />
+        <MainRightComponent />
+      </main>
+    );
+  }
+}
+
+export default MainComponent;
